integrate toggling in show/hide devconsole when enabled in game session
the control combination key to trigger showing devconsole is fixed to be
ctrl+alt+c key. It used monospaced baked-in font called "hankaku_font10"
which is used in menu screen to eliminate a need to add additional font
resource to the module.

Logic code is abstracted into its separate devconsole.c/.h file.

Note: currently implemented a testing 2 commands "cmd1" and "cmd2" for
purpose of debugging.

The step to hook this up is to
1. created and define dev.txt file with a line of "console 1"
2. play or load the game
3. press combination key of ctrl+alt+c key
4. try to type in "cmd1" or "cmd2" then enter, you will see a testing
short text as result